Generify AvatarSettings and Customization, move story data from lobby to resourcedata.
