+++ Game/RainMeadow.CustomizationHooks.cs
@@ -2,6 +2,7 @@
 using MonoMod.Cil;
 using System;
 using System.Linq;
+using System.Runtime.CompilerServices;
 using UnityEngine;
 
 namespace RainMeadow
@@ -9,6 +10,8 @@ namespace RainMeadow
     // Support character customization (WIP)
     public partial class RainMeadow
     {
+        public static ConditionalWeakTable<Creature, AvatarSettings.AvatarCustomization> creatureCustomizations = new();
+
         public void CustomizationHooks()
         {
             IL.PlayerGraphics.ApplyPalette += PlayerGraphics_ApplyPalette;
@@ -39,7 +42,7 @@ namespace RainMeadow
                     c.Emit(OpCodes.Ldarg_0);
                     c.Emit(OpCodes.Ldloca, colorVar.Index);
                     c.EmitDelegate((PlayerGraphics self, ref Color originalEyeColor) => {
-                        if (MeadowCustomization.creatureCustomizations.TryGetValue(self.player, out var customization))
+                        if (RainMeadow.creatureCustomizations.TryGetValue(self.player, out var customization))
                         {
                             customization.ModifyEyeColor(ref originalEyeColor);
                         }
@@ -65,7 +68,7 @@ namespace RainMeadow
                 c.Emit(OpCodes.Ldarg_0);
                 c.Emit(OpCodes.Ldloca, 0);
                 c.EmitDelegate((PlayerGraphics self, ref Color originalBodyColor) => {
-                    if (MeadowCustomization.creatureCustomizations.TryGetValue(self.player, out var customization))
+                    if (RainMeadow.creatureCustomizations.TryGetValue(self.player, out var customization))
                     {
                         customization.ModifyBodyColor(ref originalBodyColor);
                     }
