+++ GameModes/StoryGameMode.cs
@@ -1,16 +1,22 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 
 namespace RainMeadow
 {
     public class StoryGameMode : OnlineGameMode
     {
         public string myDenPos = "SU_C04";
-        public ushort karma;
         public List<ushort> readyForWinPlayers = new List<ushort>();
+        // these are synced by StoryLobbyData
+        public ushort karma;
         public bool didStartGame = false;
         public bool didStartCycle = false;
+
+        public StoryAvatarSettings storyAvatarSettings => avatarSettings as StoryAvatarSettings;
+
         public StoryGameMode(Lobby lobby) : base(lobby)
         {
+
         }
         public override bool AllowedInMode(PlacedObject item)
         {
@@ -42,8 +48,10 @@ namespace RainMeadow
             return true;
         }
 
-        public override bool PlayerCanOwnResource(OnlinePlayer from, OnlineResource onlineResource) {
-            if (onlineResource is WorldSession) {
+        public override bool PlayerCanOwnResource(OnlinePlayer from, OnlineResource onlineResource)
+        {
+            if (onlineResource is WorldSession)
+            {
                 return lobby.owner == from;
             }
             return true;
@@ -51,9 +59,34 @@ namespace RainMeadow
 
         public override void LobbyReadyCheck()
         {
-            if (OnlineManager.lobby.isOwner && !(OnlineManager.lobby.gameMode as StoryGameMode).didStartGame) {
-                (OnlineManager.lobby.gameMode as StoryGameMode).didStartGame = true;
+            if (lobby.isOwner && !didStartGame)
+            {
+                didStartGame = true;
             }
+        }
+
+        internal override void PlayerLeftLobby(OnlinePlayer player)
+        {
+            base.PlayerLeftLobby(player);
+            if (player == lobby.owner)
+            {
+                OnlineManager.instance.manager.RequestMainProcessSwitch(ProcessManager.ProcessID.MainMenu);
+            }
+        }
+
+        internal override void ResourceAvailable(OnlineResource onlineResource)
+        {
+            base.ResourceAvailable(onlineResource);
+            if(onlineResource is Lobby lobby)
+            {
+                lobby.AddData<StoryLobbyData>(true);
+            }
+        }
+
+        internal override void LobbyTick(uint tick)
+        {
+            base.LobbyTick(tick);
+            readyForWinPlayers = lobby.entities.Values.Where(e => e.entity is StoryAvatarSettings sas && sas.readyForWin).Select(e => e.entity.owner.inLobbyId).ToList();
         }
     }
 }
