+++ Meadow/MeadowWorldData.cs
@@ -1,28 +1,34 @@
-﻿namespace RainMeadow
+﻿using System;
+
+namespace RainMeadow
 {
     internal class MeadowWorldData : OnlineResource.ResourceData
     {
         internal ushort spawnedItems;
 
-        internal override OnlineResource.ResourceDataState MakeState(OnlineResource inResource)
+        public MeadowWorldData(OnlineResource resource) : base(resource) { }
+
+        internal override ResourceDataState MakeState()
         {
-            return new MeadowWorldState(this);
+            return new State(this);
         }
 
-        internal class MeadowWorldState : OnlineResource.ResourceDataState
+        internal class State : ResourceDataState
         {
             [OnlineField]
             ushort spawnedItems;
 
-            public MeadowWorldState() { }
-            public MeadowWorldState(MeadowWorldData meadowWorldData)
+            public State() { }
+            public State(MeadowWorldData meadowWorldData)
             {
                 spawnedItems = meadowWorldData.spawnedItems;
             }
 
-            internal override void ReadTo(OnlineResource onlineResource)
+            internal override Type GetDataType() => typeof(MeadowWorldData);
+
+            internal override void ReadTo(OnlineResource.ResourceData data)
             {
-                onlineResource.GetData<MeadowWorldData>(true).spawnedItems = spawnedItems;
+                (data as MeadowWorldData).spawnedItems = spawnedItems;
             }
         }
     }
